+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- fix(sourcemaps): Re-read package.json after CLI install (#489)
 - fix(nextjs): Set created test route handler to always be dynamic (#486)
 
 ## 3.16.1